Added status bluetooth of NRF52
