+++ src/nimble/BluetoothUtil.h
@@ -29,19 +29,3 @@ int chr_readwrite32le(uint32_t *v, struct ble_gatt_access_ctxt *ctxt);
  */
 int chr_readwrite8(uint8_t *v, size_t vlen, struct ble_gatt_access_ctxt *ctxt);
 
-enum StateBT 
-{
-	CONNECT,
-	DISCONNECT,
-	CONN_UPDATE,
-	ADV_COMPLETE,
-	ENC_CHANGE,
-	SUBSCRIBE,
-	MTU,
-	REPEAT_PAIRING,
-	PASSKEY_ACTION,
-	ON,
-	OFF
-};
-
-extern StateBT stateBT;
